+++ source/web/src/JSDraw.MolHandler.ts
@@ -1,20 +1,24 @@
-import type {JSDraw2ModuleType} from './types/jsdraw2';
+import type {IDrawOptions, IEditorOptions, JSDraw2ModuleType} from './types/jsdraw2';
 import type {ScilModuleType} from './types';
 
 import type {IMolHandler, MolHandlerOptions} from './types/mol-handler';
 
 import {defaultMolHandlerOptions} from './types/mol-handler';
 import {Mol} from './Mol';
+import type {OrgType} from './types/org';
 
 declare const JSDraw2: JSDraw2ModuleType;
 declare const scil: ScilModuleType;
+declare const org: OrgType<any, IEditorOptions>;
 
-export class MolHandler<TBio> implements IMolHandler<TBio> {
+export class MolHandler<TBio, TEditorOptions extends IEditorOptions>
+  implements IMolHandler<TBio, TEditorOptions> {
   public readonly T: string;
-  private readonly options: MolHandlerOptions;
+
 
   public readonly bondlength: number;
   public readonly m: Mol<TBio>;
+  public readonly options: Partial<TEditorOptions>;
 
   constructor(options?: MolHandlerOptions) {
     this.T = 'MOLHANDLER';