+++ source/web/src/types/mol-handler.ts
@@ -1,4 +1,5 @@
 import type {Mol} from '../Mol';
+import type {IEditorOptions} from './jsdraw2';
 
 export const defaultMolHandlerOptions = new class {
   showimplicithydrogens: boolean = true;
@@ -7,7 +8,8 @@ export const defaultMolHandlerOptions = new class {
 export type MolHandlerOptions = typeof defaultMolHandlerOptions;
 
 /** Molecular handler for parser plugin. */
-export interface IMolHandler<TBio> {
+export interface IMolHandler<TBio, TEditorOptions extends IEditorOptions> {
   get bondlength(): number;
   get m(): Mol<TBio>;
+  get options(): Partial<TEditorOptions>;
 }